CARDS-1761: Adapt React Date / Time pickers Moved DateTimeUtilities to commons package
