+++ modules/patient-portal/src/main/frontend/src/patient-portal/QuestionnaireSet.jsx
@@ -44,7 +44,7 @@ import { DateTime } from "luxon";
 
 import Form from "../questionnaire/Form.jsx";
 import Header from "./Header.jsx";
-import DateTimeUtilities from "../questionnaire/DateTimeUtilities";
+import DateTimeUtilities from "../components/DateTimeUtilities";
 import FormattedText from "../components/FormattedText.jsx";
 import { ENTRY_TYPES } from "../questionnaire/FormEntry.jsx"
